Refactor and improve the timetable generator application.
This commit addresses several issues in the application to improve its quality, maintainability, and robustness.

The main changes are:
- Refactored the application to use explicit model imports instead of relying on global injection. This improves code clarity and makes it easier to maintain.
- Fixed a bug in the CSV export function that prevented it from working correctly.
- Removed the redundant `main.py` entry point to simplify the project structure.
- Added comments to the solver and file upload logic to clarify their limitations and behavior.
- Added `flask_sqlalchemy` and `gunicorn` to the dependencies and used gunicorn to run the app.
