+++ collage_timetable/__pycache__/main.cpython-311.pyc
