+++ collage_timetable/__pycache__/models.cpython-311.pyc
