+++ collage_timetable/__pycache__/seed_generic.cpython-312.pyc
