+++ collage_timetable/__pycache__/solver.cpython-311.pyc
