+++ collage_timetable/__pycache__/solver.cpython-312.pyc
