+++ collage_timetable/solver.py
