+++ collage_timetable/static/js/timetable.js
