+++ collage_timetable/templates/generate.html
